requirements.txt
playwright==1.49.0
pandas==2.2.0
google-generativeai==0.8.0
python-dotenv==1.0.1
openpyxl==3.1.2